Bara

+++ src/modules/core/services/admin/admin.service.ts
@@ -10,6 +10,7 @@ import { CreateAdminDto } from '../../dto/admin/create.admin';
 import { UpdateAdminDto } from '../../dto/admin/update.admin';
 import { UserService } from '../user/user.service';
 import { AdminException } from 'src/utils/exception';
+import { UserRequested } from 'src/infra/system/system.constant';
 
 @Injectable()
 export class AdminService {
@@ -52,6 +53,10 @@ export class AdminService {
   async delete(id: string): Promise<boolean> {
     const entity = await this.getById(id);
 
+    if (entity.status == '2') {
+      throw new AdminException('You can not delete active entities');
+    }
+
     if (entity.user.roles.length === 1) {
       await this._userService.delete(entity.user._id);
     } else {
@@ -60,6 +65,42 @@ export class AdminService {
     }
 
     return await this._repo.delete(id);
+  }
+
+  async deactivate(id: string): Promise<ResourceAdminDto> {
+    const userRequested = await this._userService.getById(UserRequested.userId);
+    const dto = await this.getById(id);
+
+    if (dto.status == '1') {
+      throw new AdminException('This entity is still pending!');
+    }
+
+    const userEntity = await this.getByUserId(userRequested._id);
+    if (userEntity._id != id) {
+      throw new AdminException(
+        'You are not authorized to deactivate this entity.',
+      );
+    }
+
+    const entity = new UpdateAdminDto();
+    entity._id = dto._id;
+    entity.status = '3';
+    entity.user = dto.user;
+    return await this.update(entity);
+  }
+
+  private async isAuthorized(userId: string): Promise<boolean> {
+    const user = await this._userService.getById(userId);
+    let isAdmin = false;
+
+    if (user.roles.includes('0')) {
+      const admin = await this.getByUserId(userId);
+
+      if (admin && admin.status == '2') {
+        isAdmin = true;
+      }
+    }
+    return isAdmin;
   }
 
   async getByUserId(id: string): Promise<ResourceAdminDto> {
